Séparation en différents fichier et ajout de classe
