+++ selectionCourt.py
@@ -1,0 +1,15 @@
+## Selection du court
+addressCourt1 = "169.254.74.221"
+addressCourt2 = "169.254.74.224"
+addressCourt3 = "169.254.74.223"
+addressCourt4 = ""
+
+
+def listeCourt() -> None:
+    ## global Bouton_de_la_liste_des_cours
+    ## OU
+    ## global bouton_court1 etc
+    listeCourt = []
+    listeCourt.append(addressCourt1)
+    listeCourt.append(addressCourt2)
+    listeCourt.append(addressCourt3)
